Extract PrivateKeySource and FilePrivateKeySource
