+++ networking/p2p/src/main/java/tech/pegasys/teku/networking/p2p/network/config/PrivateKeySource.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright Consensys Software Inc., 2024
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.networking.p2p.network.config;
+
+import org.apache.tuweni.bytes.Bytes;
+
+public interface PrivateKeySource {
+  Bytes getOrGeneratePrivateKeyBytes();
+}
