+++ teku/src/test/java/tech/pegasys/teku/cli/options/P2POptionsTest.java
@@ -26,8 +26,8 @@ import tech.pegasys.teku.config.TekuConfiguration;
 import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 import tech.pegasys.teku.networking.eth2.P2PConfig;
 import tech.pegasys.teku.networking.p2p.discovery.DiscoveryConfig;
+import tech.pegasys.teku.networking.p2p.network.config.FilePrivateKeySource;
 import tech.pegasys.teku.networking.p2p.network.config.NetworkConfig;
-import tech.pegasys.teku.networking.p2p.network.config.NetworkConfig.FilePrivateKeySource;
 
 public class P2POptionsTest extends AbstractBeaconNodeCommandTest {
